requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
mysqlclient==2.2.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqladmin==0.16.1
pymysql==1.1.0
email-validator==2.1.1
itsdangerous==2.2.0
alembic==1.13.1
bcrypt==3.2.0
boto3==1.34.140
apscheduler==3.10.4
pytz==2024.1
requests
prometheus-fastapi-instrumentator
celery==5.4.0
flower==2.0.1
numpy
torch
